+++ app/Http/Controllers/Recruiter/NotificationController.php
@@ -14,6 +14,8 @@ class NotificationController extends Controller
 {
     public function index(Request $request)
     {
+        $this->authorize('viewAny', UserNotification::class);
+
         $user = Auth::user();
         
         // Build query
@@ -210,9 +212,8 @@ class NotificationController extends Controller
     public function markAsRead($id)
     {
         $user = Auth::user();
-        $userNotification = UserNotification::where('id', $id)
-            ->where('user_id', $user->id)
-            ->first();
+        $userNotification = UserNotification::findOrFail($id);
+        $this->authorize('update', $userNotification);
             
         if ($userNotification) {
             $userNotification->update([
@@ -234,6 +235,8 @@ class NotificationController extends Controller
     
     public function markAllAsRead()
     {
+        $this->authorize('viewAny', UserNotification::class);
+
         $user = Auth::user();
         UserNotification::where('user_id', $user->id)
             ->where('is_read', false)
@@ -251,9 +254,8 @@ class NotificationController extends Controller
     public function destroy($id)
     {
         $user = Auth::user();
-        $userNotification = UserNotification::where('id', $id)
-            ->where('user_id', $user->id)
-            ->first();
+        $userNotification = UserNotification::findOrFail($id);
+        $this->authorize('delete', $userNotification);
             
         if ($userNotification) {
             $userNotification->delete();
@@ -272,6 +274,8 @@ class NotificationController extends Controller
     
     public function destroyAll()
     {
+        $this->authorize('viewAny', UserNotification::class);
+
         $user = Auth::user();
         UserNotification::where('user_id', $user->id)->delete();
